requirements.txt
fastapi
uvicorn
qdrant-client
httpx
python-dotenv
pyyaml
tqdm
transformers
sentence-transformers
